feat: add t_map and main; main test main to test folder
